Clean up weak references of functions periodically as well as when the context is disposed
The periodic cleanup runs every time the number of weak references to functions exceeds 100 and removes all elements that are either flagged as not alive (the GC collected them already), or whose mFuncHandle is null (those that are collected by the GC but not yet flagged). The latter results in a more aggressive cleanup. For this to work we make mFuncHandle internal rather than private.

Without this cleanup every weak reference takes up 24 byte (as measured by the Visual Studio profiler) leading to roundabout 50 MB / 100,000 functions. Note that the JavascriptFunction objects needs to be disposed after usage client side, if you want to profit from the aggressive cleanup.

This also adds a regression test.