Fix crash when converting DateTime to V8 Date if JSON.stringify is called on a recursive data structure
When calling JSON.stringify on a recursive data structure that contains a DateTime object we experienced a crash
because of the call to ToLocalChecked in the return statement of JavascriptInterop::ConvertDateTimeToV8.
The crash happened because ToLocalChecked sends us to FatalErrorCallback (JavascriptContext.cpp) which then terminates the process.
However, there is already an exception pending when we reach that ToLocalChecked call because the maximum call stack size
has been exceeded while stringifying. This is the actual reason why we get an empty MaybeLocal in the first place which
then leads to the crash. Interestingly, this *only* seems to happen for the case of the DateTime conversion.