Catch .NET exceptions that might occur during a MoveNext call inside an iterator, schedule them as proper JS exceptions, and return an IteratorResult object indicating the end of the iterator
If we don't catch and schedule here the .NET exception leads to an immediate termination of the script execution with no way of catching the error in JS land. This can be especially bad if the JS code does manual resource management (e.g. closing a file handle) in a finally block that is then never executed.